requirements.txt
certifi==2018.8.13
chardet==3.0.4
defusedxml==0.5.0
Django==2.1
django-allauth==0.36.0
django-crispy-forms==1.7.2
django-environ==0.4.5
djangorestframework==3.8.2
idna==2.7
oauthlib==2.1.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
urllib3==1.23
